+++ blindoracle_sdk/__init__.py
@@ -18,11 +18,17 @@ Usage:
 """
 
 from blindoracle_sdk.client import BlindOracleClient
+from blindoracle_sdk.aio import AsyncBlindOracleClient
+from blindoracle_sdk.markets import Market
 from blindoracle_sdk.exceptions import (
     BlindOracleError,
     AuthenticationError,
     RateLimitError,
     MarketNotFoundError,
+    PaymentRequiredError,
+    ValidationError,
+    PassportRequiredError,
+    CredentialNotFoundError,
 )
 from blindoracle_sdk.audit import AuditAPI, AuditAttestation, verify_inclusion, verify_anchor
 from blindoracle_sdk.privacy import PrivacyAPI, DISCLOSURE_MODES, ZK_CLAIM_TYPES
@@ -35,17 +41,23 @@ from blindoracle_sdk.delegation import (
     DELEGATION_KIND,
 )
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 __author__ = "Craig Brown"
 __email__ = "craigmbrown@gmail.com"
 __url__ = "https://craigmbrown.com/blindoracle"
 
 __all__ = [
     "BlindOracleClient",
+    "AsyncBlindOracleClient",
+    "Market",
     "BlindOracleError",
     "AuthenticationError",
     "RateLimitError",
     "MarketNotFoundError",
+    "PaymentRequiredError",
+    "ValidationError",
+    "PassportRequiredError",
+    "CredentialNotFoundError",
     "AuditAPI",
     "AuditAttestation",
     "verify_inclusion",